Merges the modules cardiniality & relationship_manager into relationship
