+++ doc/source/module_documentation.rst
@@ -4,35 +4,31 @@ Modules documentation
 
 Core
 ----
+
 .. automodule:: neomodel.core
   :members:
 
+
 Properties
 ----------
+
 .. automodule:: neomodel.properties
   :members:
   :show-inheritance:
 
+
 Relationships
 -------------
+
 .. automodule:: neomodel.relationship
   :members:
   :show-inheritance:
 
-.. automodule:: neomodel.relationship_manager
-  :members:
-  :show-inheritance:
-
-.. automodule:: neomodel.cardinality
-  :members:
-  :show-inheritance:
 
 Match
 -----
+
 .. module:: neomodel.match
-.. autoclass:: neomodel.match.BaseSet
-   :members:
-   :undoc-members:
 .. autoclass:: neomodel.match.NodeSet
    :members:
    :undoc-members:
